Feature and refactor update.

+++ implicit_solvent_ddm/postTreatment.py
@@ -3,21 +3,14 @@ class that will parse in pandas dataframe for mbar analysis
 """
 import os
 import re
-from ast import parse
-from cProfile import run
-from itertools import chain
-from numbers import Complex
-from re import L
-from typing import List
+
 
 import pandas as pd
 from toil.job import Job
-
-import implicit_solvent_ddm.pandasmbar as pdmbar
 from implicit_solvent_ddm.get_dirstruct import Dirstruct
 from implicit_solvent_ddm.mdout import min_to_dataframe
+
 from implicit_solvent_ddm.restraints import RestraintMaker
-from implicit_solvent_ddm.simulations import Simulation
 
 WORKDIR = os.getcwd()
 
@@ -26,187 +19,205 @@ BOLTZMAN = 1.380658e-23
 JOULES_PER_KCAL = 4184
 
 
-class PostTreatment(Job):
+class ConsolidateData(Job):
     def __init__(
         self,
-        simulation_data: List[pd.DataFrame],
-        temp: float,
-        system: str,
-        max_conformation_force: float,
-        max_orientational_force=None,
-    ) -> None:
-        super().__init__()
-        self.simulations_data = simulation_data
-        self.temp = temp
-        self.system = system
+        complex_adative_run,
+        ligand_adaptive_run,
+        receptor_adaptive_run,
+        flat_botton_run,
+        temperature: float,
+        max_conformation_force,
+        max_orientational_force,
+        boresch_df: RestraintMaker,
+        working_path,
+        complex_filename,
+        ligand_filename,
+        receptor_filename,
+    ):
+        Job.__init__(self, memory="2G", cores=2, disk="3G")
+        self.temp = temperature
+        self.complex_adative_run = complex_adative_run
+        self.receptor_adaptive_run = receptor_adaptive_run
+        self.ligand_adaptive_run = ligand_adaptive_run
+        self.flat_botton_run = flat_botton_run
         self.max_con_force = str(max_conformation_force)
         self.max_orien_force = str(max_orientational_force)
-        self._kcals_per_Kt()
+        self.boresch = boresch_df
+        self.complex_name = re.sub(r"\..*", "", os.path.basename(complex_filename))
+        self.ligand_name = re.sub(r"\..*", "", os.path.basename(ligand_filename))
+        self.receptor_name = re.sub(r"\..*", "", os.path.basename(receptor_filename))
+        self.working_path = working_path
 
-    def _kcals_per_Kt(self):
-        self.kcals_per_Kt = ((BOLTZMAN * (AVAGADRO)) / JOULES_PER_KCAL) * self.temp
+    @property
+    def complex_mbar_formatted_df(self) -> pd.DataFrame:
+        return self.complex_adative_run[1] * self.kcals_per_Kt
 
-    def _load_dfs(self):
-        self.df = pd.concat(self.simulations_data, axis=0, ignore_index=True)
+    @property
+    def complex_fe(self) -> pd.DataFrame:
+        return self.complex_adative_run[0][0] * self.kcals_per_Kt
 
-        self.name = self.df["solute"].iloc[0]
+    @property
+    def complex_error(self) -> pd.DataFrame:
+        return self.complex_adative_run[0][1] * self.kcals_per_Kt
 
-    def _create_MBAR_format(self):
-        self.df = self.df.set_index(
-            [
-                "solute",
-                "parm_state",
-                "extdiel",
-                "charge",
-                "parm_restraints",
-                "traj_state",
-                "traj_extdiel",
-                "traj_charge",
-                "traj_restraints",
-                "Frames",
-            ],
-            drop=True,
+    @property
+    def ligand_mbar_formatted_df(self) -> pd.DataFrame:
+        return self.ligand_adaptive_run[1] * self.kcals_per_Kt
+
+    @property
+    def ligand_fe(self) -> pd.DataFrame:
+        return self.ligand_adaptive_run[0][0] * self.kcals_per_Kt
+
+    @property
+    def ligand_error(self) -> pd.DataFrame:
+        return self.ligand_adaptive_run[0][1] * self.kcals_per_Kt
+
+    @property
+    def receptor_mbar_formatted_df(self) -> pd.DataFrame:
+        return self.receptor_adaptive_run[1] * self.kcals_per_Kt
+
+    @property
+    def receptor_fe(self):
+        return self.receptor_adaptive_run[0][0] * self.kcals_per_Kt
+
+    @property
+    def receptor_error(self) -> pd.DataFrame:
+        return self.receptor_adaptive_run[0][1] * self.kcals_per_Kt
+
+    @property
+    def flat_bottom_fe(self):
+        return self.flat_botton_run[0][0] * self.kcals_per_Kt
+
+    @property
+    def kcals_per_Kt(self):
+        return ((BOLTZMAN * (AVAGADRO)) / JOULES_PER_KCAL) * self.temp
+
+    @property
+    def _get_ligand_deltaG(self):
+        return self.ligand_fe.loc[
+            ("endstate", "78.5", "1.0", "0.0"),
+            [("electrostatics", "0.0", "0.0", self.max_con_force)],
+        ].values[0]
+
+    @property
+    def _get_receptor_deltaG(self):
+        return self.receptor_fe.loc[
+            ("endstate", "78.5", "1.0", "0.0"),
+            [("no_gb", "0.0", "1.0", self.max_con_force)],
+        ].values[0]
+
+    @property
+    def _get_complex_deltaG(self):
+        return self.complex_fe.loc[
+            (
+                "no_interactions",
+                "0.0",
+                "0.0",
+                f"{self.max_con_force}_{self.max_orien_force}",
+            ),
+            [("endstate", "78.5", "1.0", "0.0_0.0")],
+        ].values[0]
+
+    @property
+    def _flat_bottom_contribution(self):
+        return self.flat_bottom_fe.loc[
+            (
+                ("endstate", "78.5", "1.0", "0.0_0.0"),
+                [("no_flat_bottom", "78.5", "1.0", "0.0_0.0")],
+            )
+        ].values[0]
+
+    @property
+    def _get_boresch_standard_state(self):
+        return self.boresch.boresch_deltaG["DeltaG"].values[0]
+
+    @property
+    def compute_binding_deltaG(self) -> float:
+        return (
+            self._get_complex_deltaG
+            + self._get_ligand_deltaG
+            + self._get_receptor_deltaG
+            + self._get_boresch_standard_state
+            + self._flat_bottom_contribution
         )
 
-        self.df = self.df[["ENERGY"]]
-        self.df = self.df.unstack(["parm_state", "extdiel", "charge", "parm_restraints"])  # type: ignore
-        self.df = self.df.reset_index(["Frames", "solute"], drop=True)
-        states = [_ for _ in zip(*self.df.columns)][1]
-        extdiels = [_ for _ in zip(*self.df.columns)][2]
-        charges = [_ for _ in zip(*self.df.columns)][3]
-        restraints = [_ for _ in zip(*self.df.columns)][4]
-
-        column_names = [
-            (state, extdiel, charge, restraint)
-            for state, extdiel, charge, restraint in zip(
-                states, extdiels, charges, restraints
-            )
-        ]
-
-        self.df.columns = column_names  # type: ignore
-
-        # divide by Kcal per Kt
-        self.df = self.df / self.kcals_per_Kt
-
-    def compute_binding_deltaG(
-        self, system1: float, system2: float, boresch_dG: float, free_flat_bottom: float
-    ):
-        return self.deltaG + system1 + system2 + boresch_dG + free_flat_bottom
-
     def run(self, fileStore):
-        self._load_dfs()
-        self._create_MBAR_format()
-        fileStore.logToMaster(f"self.df {self.df}")
+        output_path = os.path.join(
+            f"{self.working_path}", f".cache/{self.complex_name}"
+        )
 
-        equil_info = pdmbar.detectEquilibration(self.df)
+        if not os.path.exists(output_path):
+            os.makedirs(output_path)
+        # parse out formatted dataframe
+        self.complex_mbar_formatted_df.to_hdf(
+            f"{output_path}/{self.complex_name}_formatted.h5", key="df", mode="w"
+        )
+        self.receptor_mbar_formatted_df.to_hdf(
+            f"{output_path}/receptor_{self.receptor_name}_formatted.h5",
+            key="df",
+            mode="w",
+        )
+        self.ligand_mbar_formatted_df.to_hdf(
+            f"{output_path}/ligand_{self.ligand_name}_formatted.h5", key="df", mode="w"
+        )
 
-        df_subsampled = pdmbar.subsampleCorrelatedData(self.df, equil_info=equil_info)
+        # parse out free energies
+        self.complex_fe.to_hdf(
+            f"{output_path}/{self.complex_name}_fe.h5", key="df", mode="w"
+        )
+        self.receptor_fe.to_hdf(
+            f"{output_path}/receptor_{self.receptor_name}_fe.h5", key="df", mode="w"
+        )
+        self.ligand_fe.to_hdf(
+            f"{output_path}/ligand_{self.ligand_name}_fe.h5", key="df", mode="w"
+        )
 
-        fe, error, mbar = pdmbar.mbar(df_subsampled)
+        # parse out errors of mbar
+        self.complex_error.to_hdf(
+            f"{output_path}/{self.complex_name}_error.h5", key="df", mode="w"
+        )
+        self.receptor_error.to_hdf(
+            f"{output_path}/receptor_{self.receptor_name}_error.h5", key="df", mode="w"
+        )
+        self.ligand_error.to_hdf(
+            f"{output_path}/ligand_{self.ligand_name}_error.h5", key="df", mode="w"
+        )
 
-        fe = fe * self.kcals_per_Kt
+        # parse out boresch restraints dataframe
 
-        # then multiply by kt
-        error = error * self.kcals_per_Kt
+        self.boresch.boresch_deltaG.to_hdf(
+            f"{output_path}/boresch_{self.complex_name}.h5", key="df", mode="w"
+        )
+        deltaG_df = pd.DataFrame()
 
-        if self.system == "ligand":
-            self.deltaG = fe.loc[("endstate", "78.5", "1.0", "0.0"), [("electrostatics", "0.0", "0.0", self.max_con_force)]].values[0]  # type: ignore
+        fileStore.logToMaster(
+            f"BORESCH standard state {self._get_boresch_standard_state}\n"
+        )
 
-        elif self.system == "receptor":
-            self.deltaG = fe.loc[("endstate", "78.5", "1.0", "0.0"), [("no_gb", "0.0", "1.0", self.max_con_force)]].values[0]  # type: ignore
+        fileStore.logToMaster(f"Ligand Delta G: {self._get_ligand_deltaG}")
+        fileStore.logToMaster(f"Receptor Delta G: {self._get_receptor_deltaG}\n")
 
-        elif self.system == "free_flat_bottom":
-            self.deltaG = fe.loc[
-                (
-                    ("endstate", "78.5", "1.0", "0.0_0.0"),
-                    [("no_flat_bottom", "78.5", "1.0", "0.0_0.0")],
-                )
-            ].values[0]
+        fileStore.logToMaster(
+            f"Complex unique index keys:\n {self.complex_fe.index.unique()}\n"
+        )
 
-        # system is complex
-        else:
-            self.deltaG = fe.loc[("no_interactions", "0.0", "0.0", f"{self.max_con_force}_{self.max_orien_force}"), [("endstate", "78.5", "1.0", "0.0_0.0")]].values[0]  # type: ignore
+        fileStore.logToMaster(f"Complex Delta G: {self._get_complex_deltaG}\n")
 
-        self.fe = fe
-        self.error = error
-        self.mbar = mbar
-        return self
+        deltaG_df[f"{self.ligand_name}_endstate->no_charges"] = [
+            self._get_ligand_deltaG
+        ]
+        deltaG_df[f"{self.receptor_name}_endstate->no_gb"] = [self._get_receptor_deltaG]
+        deltaG_df["boresch_restraints"] = [self._get_boresch_standard_state]
+        deltaG_df["flat_bottom_contribution"] = [self._flat_bottom_contribution]
+        deltaG_df[f"{self.complex_name}_no-interactions->endstate"] = [
+            self._get_complex_deltaG
+        ]
+        deltaG_df["deltaG"] = [self.compute_binding_deltaG]
 
-
-def consolidate_output(
-    job,
-    ligand_system: PostTreatment,
-    receptor_system: PostTreatment,
-    complex_system: PostTreatment,
-    flat_bottom: PostTreatment,
-    boresch_df: RestraintMaker,
-):
-    output_path = os.path.join(f"{WORKDIR}", f".cache/{complex_system.name}")
-    if not os.path.exists(output_path):
-        os.makedirs(output_path)
-
-    # parse out formatted dataframe
-    complex_system.df.to_hdf(
-        f"{output_path}/{complex_system.name}_formatted.h5", key="df", mode="w"
-    )
-    receptor_system.df.to_hdf(
-        f"{output_path}/receptor_{complex_system.name}_formatted.h5", key="df", mode="w"
-    )
-    ligand_system.df.to_hdf(
-        f"{output_path}/ligand_{complex_system.name}_formatted.h5", key="df", mode="w"
-    )
-
-    # parse out free energies
-    complex_system.fe.to_hdf(
-        f"{output_path}/{complex_system.name}_fe.h5", key="df", mode="w"
-    )
-    receptor_system.fe.to_hdf(
-        f"{output_path}/receptor_{complex_system.name}_fe.h5", key="df", mode="w"
-    )
-    ligand_system.fe.to_hdf(
-        f"{output_path}/ligand_{complex_system.name}_fe.h5", key="df", mode="w"
-    )
-
-    # parse out errors of mbar
-    complex_system.error.to_hdf(
-        f"{output_path}/{complex_system.name}_error.h5", key="df", mode="w"
-    )
-    receptor_system.error.to_hdf(
-        f"{output_path}/receptor_{complex_system.name}_error.h5", key="df", mode="w"
-    )
-    ligand_system.error.to_hdf(
-        f"{output_path}/ligand_{complex_system.name}_error.h5", key="df", mode="w"
-    )
-
-    # parse out boresch restraints dataframe
-
-    boresch_df.boresch_deltaG.to_hdf(
-        f"{output_path}/boresch_{complex_system.name}.h5", key="df", mode="w"
-    )
-
-    boresch_dG = boresch_df.boresch_deltaG["DeltaG"].values[0]
-    # compute total deltaG
-    deltaG_tot = complex_system.compute_binding_deltaG(
-        system1=ligand_system.deltaG,
-        system2=receptor_system.deltaG,
-        boresch_dG=boresch_dG,
-        free_flat_bottom=flat_bottom.deltaG,
-    )  # type: ignore
-
-    deltaG_df = pd.DataFrame()
-
-    deltaG_df[f"{ligand_system.name}_endstate->no_charges"] = [ligand_system.deltaG]
-    deltaG_df[f"{receptor_system.name}_endstate->no_gb"] = [receptor_system.deltaG]
-    deltaG_df["boresch_restraints"] = [boresch_dG]
-    deltaG_df[f"{complex_system.name}_no-interactions->endstate"] = [
-        complex_system.deltaG
-    ]
-    deltaG_df["free->flat_bottom"] = flat_bottom.deltaG
-    deltaG_df["deltaG"] = [deltaG_tot]
-
-    deltaG_df.to_hdf(
-        f"{output_path}/deltaG_{complex_system.name}.h5", key="df", mode="w"
-    )
+        deltaG_df.to_hdf(
+            f"{output_path}/deltaG_{self.complex_name}.h5", key="df", mode="w"
+        )
 
 
 def create_mdout_dataframe(
@@ -220,6 +231,7 @@ def create_mdout_dataframe(
 
     mdout = f"{output_dir}/mdout"
 
+    job.log(f"List files in postprocess directory {os.listdir(output_dir)}\n")
     run_args = sim.dirStruct.fromPath2Dict(mdout)
     data = min_to_dataframe(mdout)
 
@@ -251,6 +263,7 @@ def create_mdout_dataframe(
         )
         # data.to_parquet(f"{output_dir}/simulation_mdout.zip",  compression="gzip")
 
-    os.remove(mdout)
+    if os.path.exists(mdout):
+        os.remove(mdout)
 
     return data
